Store and return counters as a map if id->value rather than a list
Summary:
As titled. For watermark counters we need to query counters
with READ_AND_CLEAR mode. This means we will have 2 lists of
counters to query for objs with counters. So the assumption
that counters are in contigous order w.r.t. the counter ids
stored in a list becomes very fragile.
Towards that end, start storing and consuming these in a map

Differential Revision: D22175029

fbshipit-source-id: 35c51769358f4ac6481acf8016d0d2a5fa6beb76