When updating a subset of counters be explicit about the counter read mode
Summary:
As titled. This is working towards the case where
some counters would need to be read with READ and others
as READ_AND_CLEAR.
When requesting update of all counters, SaiObject will do the work of
determining read mode for each counter. However when requesting a
subset put the onus on the client.
TODO: Evaluate having SaiObject figure out the modes when a subset
of counter ids are passed in as well.

Differential Revision: D22177722

fbshipit-source-id: c8514285b1925b9d8e6175d08959a784626aea6d