Be explicit for stats read mode at the API level
Summary:
Now that we distinguish b/w stats that should
just be read vs those that need to cleared on read. Have the
callers be explicit about which set of stats they are asking
for at the sai api level.
At the SaiObject level where we want to collect all stats
for a object (port, queue) have sai object do the collection
for both read and read and clear stats

Differential Revision: D22178407

fbshipit-source-id: 3ec542a2971593adb21b52417265234fc0ba5382