Wait for both port and queue byte inc in ensurePacketSent
Summary:
Port and queue stats may not increment with
the same speed. In particular, for brcm-sai using port
uncached property seems to just get us synchrnous
port stat collection. So when ensuring a packet
has been sent out wait for port and a queue on that
port to get their out bytes incremented.

Differential Revision: D22229357

fbshipit-source-id: af479c69c516b38d9b64198370f3623b42081355