Pass port descriptor to disable TTL decrements too
Summary:
As titled. On some ASICs we can't disable
on ttl decrement on nhop, but rather need to do it
at the port level. So allow for that information
to be passed through.

TODO: Split up the actual worker methods into 2
and call the right one based on ASIC capability

Reviewed By: shri-khare

Differential Revision: D22304886

fbshipit-source-id: 43d799b22d00dc3a462cffdb6c7652a4c1e2a899