Enable monitored port changes.
Summary:
(1) Allow ports in the configuration to be monitored for QCM. (so far only tests could tickle the port monitoring) Criteria is (a) that ports should be admin up (b) gports should be available in the QCM. (we are limited to 164 gports today)
(2) Allow user configured ports to take priority in monitoring. (getMonitorQcmPortList) by programming them during init. Today these are uplink ports only pushed from the cfg. Other ports are best effort programmed after these ones are.
(3) Ensure that we use 4 gports for every port (to monitor first 4 queues), but monitor 5 gports for every port which has mh-nic enabled.
(4) Modify the qcm list of ports to a map which store gportCount for each port.
(QcmMonitorPortMap)
(5) Tests. Removed explicit monitoring of ports in QCM tests (updateQcmMonitoredPortsIfNeeded) . The normal code path with cfg updates should take care of it.

Reviewed By: shri-khare

Differential Revision: D22112407

fbshipit-source-id: 700421c4a8aa012bd38e6c2b8db5ad489f440a34