Create queue manager before hostif mgr
Summary:
The latter refers to it in its constructor when
loading CPU queues via the loadQueues API.
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/fboss/agent/hw/sai/switch/SaiHostifManager.cpp?commit=f21b658172758270a62b65b4ab8d158fdf45073a&lines=343%2C349%2C356

Why did this work at all?
The call to SaiQueueManager::loadQueues is transformed to a regular
function call with the first parameter as the this pointer (this process
is common for all C++ member function). So things work until you access
a member variable in loadQueues at which point this pointer is accessed
and we get a crash. Since loadQueues was not accessing any member variables
we were fine, but one of subsequent diffs will access member variables
there.

Reviewed By: shri-khare

Differential Revision: D22494961

fbshipit-source-id: fc34fc0a949a47377861997ee828f2b3064be962