Features enum at HwSwitchEnsemble
Summary:
Till now we only needed to turn on HwSwitch level
feature knobs in tests and benchmarks. But now we desire to
turn on stats collections for benchmarks. Motivation for this
being that in SAI stats compete for the same lock with other
HwSwitch operations, so we should account for that cost. To
that end, build a features abstraction at HwSwitchEnsemble
and translate to HwSwitch features when creating HwSwitch

Differential Revision: D22500397

fbshipit-source-id: 1a9f0bfaba0dc3c0de168dbd7217cf8b430b09a7