Remove unused NeighborListenerClient.cpp
Summary:
Seems this was used as the listener part of neighbor events which we used when Bgp
was written in D (D could call into C++ code). bgpcpp has its own neighbor
watcher implementation so this is no longer needed (Thanks Jasmeet).

This file is not built by buck TARGET anyway, so remove the file and stop
building from OSS.

Reviewed By: rsunkad

Differential Revision: D22577990

fbshipit-source-id: 9faed8ec64f8879a6484c7ef33263f3689093594