Track run state at HwSwitch level and ignore noop transitions
Summary:
In my diff stack at D22511306 there was a bug where
HwSwitch switchRunStateChanged was being called twice for
transition to CONFIGURED state. This caused linkscan to be
started twice on SaiSwitch which lead to other downstream
problems (caught by test). While I could fix that bug,
a better approach seems to be to just cache this at HwSwitch
layer.

TODO: get rid of runState on HwSwitchEnsemble and perhaps
also on SwSwitch. Just ask HwSwitch for this info when needed

Reviewed By: phshaikh

Differential Revision: D22539560

fbshipit-source-id: 4530b23a6adba2921d3b0f433e870aebf4f58373