Turn on stats collection for all benchmarks(except stats collection speed)
Summary:
As titled. We need to account for this cost and can't just
assume that it does not have a effect on other operations (in BcmSwitch
it mostly doesn't since stats collections is on a thread that does
not compete for for switch lock, but that's not the case for SaiSwitch).

Reviewed By: phshaikh

Differential Revision: D22511306

fbshipit-source-id: 9618ca1b7858134025ed530352ca99ceebb80271