Fix a nasty misplaced parenthesis bug in Mac learning tests
Summary:
When verifying mac learning we want to verify
- In HW learning mode that mac was learnt/aged in HW
- In SW learning mode that mac was learnt/aged in SW

However our if condition was written so
if (((l2LearningMode == cfg::L2LearningMode::SOFTWARE &&
     wasMacLearntInSwitchState(shouldExist)) ||
    (l2LearningMode == cfg::L2LearningMode::HARDWARE)) &&
     wasMacLearntInHw(portDescr, shouldExist)) {
  return true;
}

This reads return true - if (learning mode is sw and mac was learnt in switch or
learning mode is HW) && mac was learnt in HW

This meant we were asking for MAC to be learnt in HW even for SW learning
tests. This was not the intention of test. Fix logic now

Differential Revision: D22614985

fbshipit-source-id: e18dfc6c3247dc71da0eb2bdbe600b4a63cee4df