Change Maclearning test config to use oneL3IntfTwoPortConfig with 1 port down
Summary:
Change Maclearning test config to use oneL3IntfTwoPortConfig,
with one port as down. The motivation here is as follows
- SAI doesn't provide a way to get all L2 entries. In SW learning
mode we could refer to our FDB table, but for HW learning mode the
entries just exist in HW L2 table, which I don't see an api to query.
- So the alternative is to infer this via traffic, if a MAC was learnt
on a port then if we have in vlan traffic destined to that MAC it should
only go out that port and not get flooded in VLAN.

So to that end change config to have 2 ports in the same vlan, with one
port down. During the check for MAC learning I will enable the other
port, send traffic and check stats.

Differential Revision: D22608395

fbshipit-source-id: a709c7332fa03f35d4190754e22e2d4613496c26