Determine whether MAC was learnt or not by sending traffic to learnt dstMAC
Summary:
- SAI doesn't provide a way to get all L2 entries. In SW learning
mode we could refer to our FDB table, but for HW learning mode the
entries just exist in HW L2 table, which I don't see an api to query.
- So the alternative is to infer this via traffic, if a MAC was learnt
on a port then if we have in vlan traffic destined to that MAC it should
only go out that port and not get flooded in VLAN.
To that end change the wasMacLearntInHw api to use the above approach.

TODO: Change MAC scale test to adopt the same approach

Reviewed By: shri-khare

Differential Revision: D22618334

fbshipit-source-id: 9a80cf04af632e36a59fcf755df579f065e68c8b