PDDC test changes.
Summary:
(1) Add test for PDDC mode. Check that flow learning is 0, but flow tracker is enabled. Run in loop of 10.
(2) Add test for PDDC mode with scanInterval = 0. Ensure that when we read back scan interval after sending traffic, it comes > 0.
(3) Modify current start/stop QCM test to run in loop of 10 for stress testing a bit.

Reviewed By: steinbrecher

Differential Revision: D22738734

fbshipit-source-id: 427de93f66dff61e9be609bc3b6bc5dcb161af24