Fix warm-boot issue for SFLOW when the tunnel is unresolved.
Summary:
If tunnel is unresolved, we do not store the tunnel information in the switchState today. So during warm-boot we recover without udp ports and cause an exception while initializing (as we don't think its SFLOW mirror configuration anymore since udpports are missing)

Fix it to store this information in mirror configuration as well.

We could eventually remove this field from Tunnel Cfg, but should keep reading part atleast same until new changes are released for warmboot from older versions.

Reviewed By: phshaikh

Differential Revision: D22783563

fbshipit-source-id: 49faa3d3f98801b1a286289c619261222f8ced97