Fix Sw learning for trunk test
Summary:
As titled. In our initial configs we have 2 ports enabled
and then bring down one of them before starting the MAC learning tests.
We were doing this at the start of setupHelper. However for the
trunk based tests we reapply modified initial config (to configure
trunks) which would bring the port back up. This made the test racy
in that, if the port came up before we sent packets and checked for
tests test would fail, since flooding would ensue vs if it came up
later we would be fine

Differential Revision: D22798525

fbshipit-source-id: bcc9495ef0084ca32c75befb981d322267b1546a