Add ACL with qualifier for AclStat tests
Summary:
This is towards supporting ACLs in SAI implementation.

SAI spec requires that at least one qualifier field must be specified during
ACL entry creation:
https://github.com/opencomputeproject/SAI/blob/master/inc/saiacl.h#L1395

To be consistent, we add this requirement for BCM as well - subsequent diff in
the stack would enforce it. (note similar 'match-all' effect can be achieved by
configuring ACL with qualifiers, so OK to make this change).

With that change, the stat ACLs should have some qualifier. Add it.

Differential Revision: D22809461

fbshipit-source-id: 4c5a75366afdfd6825ec4af156f13508cf47306f