Remove AclNoQualifierTest
Summary:
This is towards supporting ACLs in SAI implementation.

SAI spec requires that at least one qualifier field must be specified during
ACL entry creation:
https://github.com/opencomputeproject/SAI/blob/master/inc/saiacl.h#L1395

To be consistent, we add this requirement for BCM as well - subsequent diff in
the stack would enforce it. (note similar 'match-all' effect can be achieved by
configuring ACL with qualifiers, so OK to make this change).

With that change, AclNoQualifierTest is no longer valid remove it.

Differential Revision: D22809462

fbshipit-source-id: 953193eb2b6eb101a9c6be73e2ba2ecdba459f98