port serdes api
Summary:
as titled, this API is used to set TX settings of a port.

adding few commonly used settings in fboss  as per phy.thrift (note - not all settings are possible, some are not available in SAI spec)

sai_get_object_key fails for serdes objects, accordingly added a trait to skip invoking it on objects which dont have support yet.

Differential Revision: D22869675

fbshipit-source-id: 467ad58f866aad676d4239e474f7701661812ec9