Account for different link down behavior b.w SAI and BcmSwitch
Summary:
There is a behavior differnce b/w SAI and BcmSwitch on link down
w.r.t. to neighbor entries. In BcmSwitch, we leave the nbr entries
intact and rely on SwSwitch to prune them. In SaiSwitch we prune
the fdb and neighbor entries on in SaiSwitch itself. Account for
that in tests

Reviewed By: shri-khare

Differential Revision: D22894469

fbshipit-source-id: 4eb6cef0b0f8c18fa3f386d5f6e582930cc0dbc4