Add SAI Replayer tests to Netcastle
Summary:
Create sai_replayer_test in netcastle. The test reuses most of the components in sai_test and bcm_test_framework since steps are the same. The main difference lies in sai_replayer_framework, where it overwrites the run_test method and add the extra step of running the generated code.

Currently, we only add one test 'HwVlanTest' as the initial test. More tests will be added on diff and continuous job.

Differential Revision: D22774454

fbshipit-source-id: 00efc73319ea18aedcf2cdf3f3875b792d1658df