Allow waiting for a given time period for l2 learn events
Summary:
In the past we noticed some unexpected events
from l2 programming. For e.g. on updating classid of
a l2 entry we would get a delete/aged event on
the old entry. I want to add some tests around
this to understand the behavior so we don't get
into a ping ping situation of deletes and adds.

I am fully expecting that some ASICs will not
generate these spurious events. To that end
add a optional interval parameter to waiting
for events, which will allow for catching of
-ves as well (no event produced 10s after a
delete for e.g.)

Reviewed By: shri-khare

Differential Revision: D22935669

fbshipit-source-id: e0f725efe0e81903614d1802cedb712d0b8e0dbe