getTransceiverInfo in platform port
Summary:
provide a pure virtual method getTransceiverInfo in platform port.

this method is implemented in implementations
- WedgePort
- BcmTestPort
- SimPlatformPort
- SaiPlatformPort

a method in SaiPlatformPort checks if platform has port2TransceiverInfo cached (because it's test), and if it is, it uses that info. Else queries QSFP.

BcmTestPort ans SaiPlatformPort checks only in platform for  port2TransceiverInfo

WedgePort queries QSFP

Differential Revision: D23042936

fbshipit-source-id: 01d686247ac613a58aad6dad9ebc09d7c1a1cba6