Add notion of static, dynamic entry type to Mac entry
Summary:
In SAI we create static L2 entries which in
response to neighbor resolution. The lifetime of such
entries is tied with neighbor entry and these are
not subject to aging. This is happening purely
at a SaiSwitch level, which is not great considering
we also will program MAC entries down in SW learning
mode. So start to pull this static FDB programming
up. Roughly my idea is that if we must program FDB
entries coeval with neighbor entries, then its better
to do so explicitly through switch state.

Base diff to first be able to distinguish b.w static
and dynamic entries. Much work will be needed if we
go down this route

Reviewed By: shri-khare

Differential Revision: D22786186

fbshipit-source-id: 1674ef46c1500a45c191168b4e37f53d61292d8d