map to maintain sw port ID to sai port id
Summary:
As titled, adding a concurrent index hashmap to maintain a mapping between swswitch PortID to sai port id (object id). This is needed for Tx.

When the packet is sent from cpu by forcing the packet our on a port, sai port id has to be provided to the Tx API as per the SAI standard.

Reviewed By: phshaikh

Differential Revision: D23071385

fbshipit-source-id: 34006aac3041c50718c6958369ad14effab26fe2