Bump up the variance tolerated in WRR from 5->10%
Summary:
Tajo ASIC does some rounding if queue weights are > 63.
The algorithm there is like so
 - Assign the highest weighted queue weight 63. Call this queue's weight
 Wmax
 - For all other queues calculate weight using the formula
round(63/(Wmax/QueueWeight))

So for our Olympic QoS deployment, we get the following actual HW weights
Gold = 63
Silver = round(63/(80/15)) = 12
ECN = round(63/(80/8)) = 6
Bronze = round(63/(80/5)) = 4

This effectively gives us the following relative traffic ratios b/w queues.
- Gold queue relative ratios -
  - Silver
    Desired - 80/15 = 5.333X (gold gets 5.33X of silver),
    Actual - 63/12 = 5.25X (gold gets 5.25X of silver)
  - Bronze
    Desired - 80/5 = 16X (gold gets 16X of bronze)
    Actual - 63/12 = 15.75 (gold gets 15.75X of bronze)
  - ECN
    Desired - 80/10 = 10X (gold gets 10X of ECN)
    Actual - 63/6 = 10.5 (gold gets 10.5X of ECN)
- Silver queue relative ratios
  - Gold - (see above)
  - Bronze
    Desired - 15/5 = 3X
    Actual - 12/4 = 3X
  - ECN
    Desired = 15/8 = 1.875
    Actual = 12/6 = 2X
- ECN relative ratios
  - Gold, Silver see above
  - Bronze
    Desired - 8/5 = 1.6X
    Actual - 6.4 = 1.5X
- Bronze relative ratios
  See above.

So how does bumping up the variance tolerated help here. Our test for WRR
always compared traffic ratios relative to the highest weighted queue and
tolerated a variance of 5%. So taking the Gold ratios above, what we got
was
For Silver a band of 5.04X - 5.59X
For ECN a band of 9.5X - 10.5X
For Bronze a band of 15.2X - 16.8X

For Tajo we observed a ratio 5X:1 b/w gold and silver. Which falls into the
band of +/-5% of 5.25X. But since we were calculating variance out of our
idealized ratio, the test failed. So allow for a variance of +-10% to account
for this difference

Reviewed By: shri-khare

Differential Revision: D23095332

fbshipit-source-id: 43037ffb1a2bf29f889960a8eeaa6e9c891a1e5b