Implement non-blocking Sai logger
Summary:
Implement non-blocking Sai logger for efficient file I/O. The main idea is to have hot/cold buffers - we start by writing to the first buffer and when it is full, immediately swap two buffers so that sai replayer can keep logging in the second buffer (non-blocking). Then in the background, another thread will flush the first buffer to disk and then ready to be swapped next time.

Flag 'disable_async_logger' is added for disabling asynchronous logging and immediately write to file as we did before.

Reviewed By: shri-khare

Differential Revision: D22875624

fbshipit-source-id: 3774e6e39312520be3921c02aef325987823805f