Allow for ignoring not found errors on delete for SaiObjects
Summary:
Selectively allow for ignoring not found in HW errors
for certain objects. An example of this is FDB entries. Consider
the case where we are transforming a dynamic entry into a
static entry. In this process, its unlikely, but possible that the
dynamic entry gets aged out from HW before we apply the changeMac(
delete + add) delta. So for FDB entries we want to ignore the
not found in HW error. For reference we already do so in the
BcmSwitch world -
https://github.com/facebook/fboss/blob/master/fboss/agent/hw/bcm/BcmMacTable.cpp#L134

Differential Revision: D23081214

fbshipit-source-id: 0421cb1484460eadfa4d81f1bfd41f6921de50fb