Outline of StaticL2ForNeighborObserver class
Summary:
As titled. In SAI we create static L2/FDB entries for each
resolved neighbor. As of now we do this in the SaiSwitch, which
conflicts with the MHNIC use case where we need to assign
class ids to L2 entries. So we decided to pull these up to
SwitchState.

Reviewed By: shri-khare

Differential Revision: D23116798

fbshipit-source-id: ad0670ed2deeb1e3665684b2f017b00e75ac130d