Abstract out a VlanTableDeltaCallbackGenerator
Summary:
Idea is to have a abstraction that processes
vlan delta and then generates calls for any updates
to Arp, Ndp, Mac entries in Vlan. We need to do this
in a few places, so having a single abstraction
doing this saves on code dup.

TODO: Make LookupClassUpdater also use this abstraction

Reviewed By: shri-khare

Differential Revision: D23120861

fbshipit-source-id: fa53a8522f93f2fb28593a1dd1ce24a71ce26e6c