Adding few parameter checks for mirror cli
Summary:
The mirror cli returns a generic fboss error when invalid parameters are provided in some conditions. Added logic to return descriptive fboss messages to cli user.

1. Added check for more than one sflow mirror.

2. Added check for non sflow type ingress mirror.

3. There is an existing check to reject egress sampling to mirror and it is based on egress rate. However if user configures egress sampling without egress rate, the check will not return error in this case. Modified the check to use egressMirror_ref instead.

Also added the following restrictions to the CLI after discussing with Rohit and Parvez

1. sflow configs need to have sampling rate specified
2. ERSPAN (gre) with sampling is not supported from cli.

These checks are implemented in the cli before calling patcher. Added it validation call which is separate from  existing configure call in cli since the configure code tries to remove the mirror when configure fails and this is not needed in validation failure case.

Differential Revision: D22918486

fbshipit-source-id: b38daad95361ef802e9887d34e9f48ee890387d0