add wrapSemiFuture, rework wrapFuture api
Summary:
This adds a new api for wrapping a semi future with
additional logging. I ended up reworking this api a bit because the
way it works right now does not play well with how we short circuit
logic when a log level is not on, leading to potential segfaults if
using wrapFuture on a log for a log level that is not enabled.

By changing to just moving the log object in to the final lambda in
the continuation chains, we also can remove some code to copy out the
variables that need to be logged.

Differential Revision: D23132459

fbshipit-source-id: ce2fdc1e2147763d4b6364ff46dddd61fea86657