Fix Async logger concurrent wait test
Summary:
Fixing Async logger concurrent wait test - the test creates another thread to emulate concurrent waiting scenario.

The issue was flushing content to file happens after the buffer is swapped, and the counter is increased after the flush completes. (Forgot this is an async logger :) )

So wait for 5ms to let the flush complete and counter updated.

Reviewed By: shri-khare

Differential Revision: D23124238

fbshipit-source-id: f04ea0592d6a2d82a89906036c209ef9ae181601