Revert to older/stricter filter of nbr entries in lookupClassUpdater
Summary:
Ideally LookupClassUpdater should only care about
where the current (added, changed) neighbor entry points to. But
the current logic for change neighbor wants the old entry to
have gone over a physical port as well. Relaxing that check
leads to crashes in places where we have LAGs. So revert to
the older logic till we can handle this transition

Reviewed By: shri-khare

Differential Revision: D23270846

fbshipit-source-id: 056c457cb5f10209df35903b2c86f3dee73c3221