Protect getProgrammedState with a mutex
Summary:
Acquire mutex to guard against picking up a stale programmed
state. A state update maybe in progress. A common pattern in
tests is to get the current programmedState, make changes to
it and then call applyNewState. If a state update is in
progress when you query programmedState_, you risk undoing
the changes of ongoing state update.

Reviewed By: shri-khare

Differential Revision: D23258181

fbshipit-source-id: d14312f168b3d29884a0c19e7fd22aaeda4a45e9