Split neighborEntry add/remove/change into separate steps
Summary:
Now that we can add/remove ports and interfaces during runtime, we run into an issue when reloading the config to remove a port (which is necessary to support 200G  speed change via a coop patcher). Because we currently remove interfaces before removing neighbor entries, we crash when trying to get the interface for the entry to remove.

This diff puts neighbor entry removal before intf removal, and leaves neighbor entry add/change after intf add.

Differential Revision: D23261183

fbshipit-source-id: 5259ecb2fe2f4e8cda1ce4ed119d1535267f212e