Cleanup: Remove FLAGS_load_qcm_fw checks (1/3)
Summary:
Cleanup QCM related flags. We no longer need this check, as we load QCM FW only when QCM is enabled ... and in next step this will help cleanup bunch of configerator settings wrt this flag.

Currently this flag is used for 2 purposes -so its overloaded. (1) To load fw (2) To generate QCM cfg for the tests.
I am  removing its usage for (1) but will keep (2).

Differential Revision: D23375627

fbshipit-source-id: 79bcae9b3e06d5ec80df6485775dbec404ec36c0