qsfp_service crash when transceiver detected with unknown type on wedge100s
Summary: The qsfp_service code detects the transceiver and then creates CmisModule or SffModule based on the module type (reg 0). If it is neither of these then it does reset. Sometime the reset results in crash in CP2112::writeReadUnsafe->CP2112::intrOut->libusb_interrupt_transfer() because the libusb handler_ is not opened yet.

Differential Revision: D23403648

fbshipit-source-id: ef75421cc8f8d09a231b26ecc3aaba98f08f746d