Towards removing separate ACLv4 and ACLv6 features
Summary:
This is towards supporting ACLs in SAI implementation.

We needed to add ACLv4 and ACLv6 as separate features as Tajo ASIC did not
support enabling both together with single SAI table.
Now that the support is added, remove the separate features.

This is first diff towards that, subsequent stacked diff will build further on
this.

Differential Revision: D23416261

fbshipit-source-id: ed2ad0fc72405a65dac5530de0c0e0c101952700