Replace separate ACLv4, ACLv6 features with single ACL feature
Summary:
This is towards supporting ACLs in SAI implementation.

We needed to add ACLv4 and ACLv6 as separate features as Tajo ASIC did not
support enabling both together with single SAI table.
Now that the support is added, remove the separate features.

Once the vendor support is complete/stable, we might remove ACL featuer
altogether and always create ACL on all platforms.

Differential Revision: D23416406

fbshipit-source-id: b5edbf41649b80145a62cd06bb2424160b8fd1e8