Creating thrift endpoint for programmatic diag shell command
Summary:
In order allow for flexible testing, we need programmatic submission of diag shell command. Adding a thrift interface to do so.

This also brings the diag cmd upto the ctrl.thrift, as both bcm and sai implements the interface.

Will add the implementation for the service endpoint in the next diff.

Differential Revision: D23020199

fbshipit-source-id: d260f4761d7db44906a51bc5e125293f1d5dd25b