Handle unclean exit
Summary:
[This diff was originally authored by Zecheng (Ron) He]

As part of the robustness requirement for async_logger, it needs to write out
any remaining log in the async buffer, even in the case of unclean exit.
Dynamically allocated memory could be corrupted when the program exits and
enter the terminate handler.
Therefore, we use static memory for async buffer and related variables such
that in terminate handler, these variables are still valid and accessible. Then
we append the remaining log in buffer to the log file.
Also increase the timeout value for async logger tests, since the test now
filled up the entire buffer (409600 bytes). It will be more robust in the case
of concurrent stress runs.

Reviewed By: rsunkad

Differential Revision: D23419260

fbshipit-source-id: 218c39a41499d739fc47144c887059dbc39bc0b8