LED utils library
Summary:
Providing common utility library for wedge platforms.

Idea is to provide magic numbers, or constants, such as LED code and LED status bytes in a library. The library also provides a computation needed to find out offset in LED banks. This library will be or can be extended further with caveat. It should not depend on ANY SDK libraries i.e. it should be independent of broadcom-sdk or sai-sdk and provides only common set of functions which apply to LED programming using both APIs

Library will be used for programming LEDs in BCM, SAI and their HW Test environments.

Differential Revision: D23456229

fbshipit-source-id: 749a9f839bb6db188e2004bc8ec7c23c84df0a56