make BcmPortTable lookups thread safe
Summary:
Replace bcm port maps with concurrent hash maps to make them thread-safe
This protects against the map being modified while an iterator is being held to the map (so should fix the current issue where we can crash while iterating through the port map in the stats thread).

There still can be issues for BcmPort* objects returned by getBcmPort. If one of these pointers is accessed after the map has been modified, we could still cause a crash. Some potential solutions to this could be returning an iterator instead of a BcmPort*, or just copying the BcmPort object and returning it (though this latter idea only would work if the BcmPort is read only).

Differential Revision: D21819829

fbshipit-source-id: c6222cd886ababd34a680c43f6b22f92b58eec49