Drop support for unused OutDropReasons attr in debug counters
Summary:
I added this as a nice to have, but turns out that BRCM-SAI
does not support set, get on this attr, so we fail on WB. There was
another but (fixed in stacked diff) which was causing us to recreate
debug counters on warm boot, which was hiding this this bug.

Drop this, since we don't use this in prod, nor do we test for it.

Differential Revision: D23722269

fbshipit-source-id: 7a70529c62eac15a3dc48302e7b7953aaf2838f1