Put debug counter object in its store, else we will recreate on each WB
Summary:
By not putting the debug counter object in store, we were actually
recreating debug counter every time and leaking the previous debug
counter in HW. Since not putting this in store meant we never serialized
the adapter key for this counter and thus never found it on WB. We then
simply proceeded on creating a new one. Since there are only 3 free cstat
counters, we ran out of resources on the 4th WB.

This is where are true get object key based solution is superior to one
where we serialize the adapter keys (known limitation).

TODO: Make the SaiObject constructors private and force construction
through store only

Differential Revision: D23722270

fbshipit-source-id: 18b344cc876970116a82538728cd134f1a4e4d4f