fix configrePortGroup to remove subsumed ports
Summary:
configurePortGroup returns immediately when a port is not in the config.

For eg,

When we are trying to configure lets say four ports in a group - port 1 through 4 with speed as 50G. Port 1 and Port 3 should be part of the config and port 2 and port 4 wil be removed/disabled based on the platform. Without the fix, it was returning before configuring port 3 to 50G mode and port 3 and port 4 was set to disabled which is incorrect.

Differential Revision: D23715686

fbshipit-source-id: 7697d543836fcbb47d5cc3debb5e8818e7216c06