Support SAI style agent config
Summary: It used to be that we have a static number of port to transceiver mapping for each platform. In SAI config, we have only the ports that will be enabled in the config. We do not keep subsumed ports. Qsfp-service needs to know how many ports should it expect. Thus we learn it from the config.

Reviewed By: tfangit

Differential Revision: D23362681

fbshipit-source-id: dc58646f0085c099da09003778dfcc1b8b5af944