Don't subject resource stats collection to logical short-circuting
Summary:
We were using the following expression to compute if any
of the resource stats were stale
```
 *stats->hw_table_stats_stale_ref() =
 !(refreshHwStatusStats(stats) && refreshLPMStats(stats) &&
  refreshFPStats(stats));
```
The downside of this is that with AND short circuiting the first
stat collection function that could not collect all stats (due to
say lack of support on the platform) would cause subsequent stats
collection to be skipped. This is not the model we want for
resource stats collection.

Differential Revision: D23686135

fbshipit-source-id: ee462f25d6aecbfa0c8acb4f5d99a9ee4e70b502