Don't remove FDB entries gratuitously
Summary:
For any change in FDB entry we were doing a remove and add.
This is not the right thing to do and would disrupt l2, l3 traffic
on (say) just a metadata update. L2 is obvious, but why l3?, the
reason for this is that in SAI the port->Neighbor/Nexthop mapping is
not present in neighbor/nexthop entries, but rather is derived from
FDB entries. So if you remove the FDB entry, neighbor corresponding
to that MAC also becomes unreachable. Transitively, next hop corresponding
to that MAC also becomes unreachable.

Further the above approach was also triggering a race in BRCM-SAI.
Consider the sequence of - FBD entry for Mac M getting programmed
w/o metadata and then being updated to have metadata due to observer
noticing that this is in a MHNIC environment. With the previous
approach, the metadata update would trigger a del and add of FDB entry.
When the delete happens
i) BRCM h/w generates a callback for l2 entry delete
ii) BRCM-SAI notices this callback and marks the egress corresponding to
this MAC as set to discard.

Now there is race, if ii) comes before the MAC entry add we will get
a egress set to discard forever (or until the next neighbor flap).

Reviewed By: phshaikh

Differential Revision: D23747410

fbshipit-source-id: 5631620ec36f42f2f831457e0c108eee4602886d