Make scheduler id part of queue create attrs
Summary:
As titled. Without this we would not load
it as part of store reload during warm boot. In
this case luckily since its just a reference,
we don't leak any resources. In that, the scheduler
store would still reload it and we would get that
obj id from it. But its still not a good model.

Differential Revision: D23776266

fbshipit-source-id: b565b11b8debc0ea1645ee78025c49b669f3ce79