claim managed next hop before creating a route
Summary:
Existing logic "assumes" route next hop does not exist and ends up creating a route first to CPU and then creates a managed next hop which redirects route to resolved neighbor.

This is great until warmboot :)

Consider a warm boot. During warm boot, a route is already present in ASIC/SDK. Routes' next hop also (possibly) exists (if it is resolved). The existing logic ends up redirecting route to CPU and then sets it back to next hop right away.

This kinds of causes disruption during warm boot.

To fix this, first claim the next hop and if it is resolved, use its id in create attributes of route.

This ensures that route's state is retained across warmboot

Differential Revision: D23798906

fbshipit-source-id: 4e470052d248cedf4e5960f4007b9197291fafb1