rename Yamp/Elbert fpga -> system
Summary:
## Renamed Yamp fpga classes:

>YampBaseFpga -> YampBaseSystem
YampFpga -> YampSystem
YampPimFpga -> YampPimSystemContainer
ElbertFpga -> ElbertSystem
ElbertPimFpga -> ElbertPimSystemContainer

## Motivation:
We want to generalize the current fpga classes to be transport agnostic. That is, in the future, registers that we're using may be accessible from different busses (ex SPI bus, i2c). In this way, we can bring together these apis and the callers do not need know about the transport protocol.

## Next steps:
We want all XXXSystems to have to following structure:

- YampSystem
  - Holds YampPimSystemContainers
    - Each Container should only be responsible to hold various controllers and exposes APIs to get each controller (for example Qsfp, Led, Mdio)

With this structure we can create pim containers based on pim type and create controllers based on transport type (not controllers all write through HwMemoryRegion which in its own is transport agnostic)

Yamp and Elbert are close to this model already, YampPimSystemContainer mainly holds various controllers and external users go through getXXXController to talk to the hw. Minipack/Fuji/Wedge400 still need work to get to this model

Reviewed By: tfangit

Differential Revision: D23711715

fbshipit-source-id: 73e814ce8b64272fc1d59340adfa92afeeae1b36