Add a test for exercising MAC move and neighbor prog interation
Summary:
As titled. As explained in the comment this exposes
a race in brcm-sai code which is being worked in CS00011145260.
Typically MAC moves occur with server moves and its unlikely
that this will be done in a time < our neighbor entry
expiry time. So I added a remove neighbor in between
move. However will chase this down with in CSP

Differential Revision: D23835803

fbshipit-source-id: 4b7cfb960b09c0fc98bdcefadb7cca3c4a5aed41