force BcmPort destruction when removed from BcmPortTable
Summary:
explicitly destroy bcmPort objects when removed from the port table

-----------
Full reason for this:
folly's concurrentHashMap doesn't destroy erased objects until a certain memory threshold is reached. The bcmPort destructor handles disabling mirroring on bcmPorts. If we don't call the destructor before updating the state with a mirror change, then the mirror will still be in use when we try to remove it and the agent will crash.

Differential Revision: D23946536

fbshipit-source-id: c96072482288bc8025fb5a2893bda85e3d0be981