Handle mismatched HW/SW trunk state scenario during warmboot init
Summary: If a trunk member goes down, link scan thread will remove member from trunk and post link down event. LACP is a listener for port oper state and mark the member as disabled. However there is a small probability that warm boot shut down happens before lacp marks member as disabled. In this case, there will be a mismatch in trunk membership in SW and HW during warm boot init. The solution is to add member to trunk when mismatch is detected and let linkUp/DownHwLocked call at end of warm boot init fix the state. The approach is similar to what is current done for ECMP members.

Differential Revision: D23936563

fbshipit-source-id: bf3f8d3e22c0a2e440f2ab9eb03dd0bdfd9c2447