Optimizations in tx packet allocation
Summary:
2 optimizations
- Inline the function (this did not get me much improvement)
- Use createCombined vs createSeparate for IOBuf. This saves
a memory allocation and gave me about a 5% speedup in tx
benchmark. This is not huge, but saving a memory allocation
can't hurt.

Note that since in SAI, we don't directly allocate DMA
member but rather just allocate on the heap and send the
tx buffer over, there will always be a perf differential.
In benchmarks this comes out to be about 25-30%. For now
I am inclined to live with it. Maybe during PVT we
can propose extensions to sai spec for packet allocation
and deallocation functions on hostif to improve
this performance

Differential Revision: D24091537

fbshipit-source-id: 2aa6b863e1b671aac1068e73b274c64440f53b84