add agent thrift api to get platform mapping
Summary:
To be able to query the current platform info from other tools (link_diag/netgram speed change workflow), we need the agent to expose its computed platform mapping.

We query this from the agent at runtime instead of compiling against the platformMapping json, since we can have a different platform mapping depending on what types of PIMs are in the chassis.

Differential Revision: D21636652

fbshipit-source-id: 3ca230ea9a80e0575ab5a5256b15006da1b4d294