collectXUnsafe to collectX
Summary:
Migration from Future-returning executor-erasing collectX forms to
SemiFuture-returning forms, that are less risky in particular with coroutines.

Earlier changes added collectXSemiFuture and collectXUnsafe as a migration
path. We then migrated collectX callsites to collectXSemiFuture or
collectXUnsafe and switched the implementation of collectX to the SemiFuture
form.

This diff renames uses of collectXUnsafe to collectX, where it is known to be
equivalent.

Differential Revision: D20989416

fbshipit-source-id: 01ff5c09c84f86f465cf75a82e4d1386d0ae6b7a