Introduce Ingress ACL Switch Attribute
Summary:
This is towards supporting ACLs in SAI implementation.

Today, FBOSS implementation creates an ACL table and ACL table group with bind
point type list that includes Switch.

However, that is not enough: the ACL table/group must be tied with specific
switch/port etc. instance.

SAI provides a switch attribute for doing that. Use it.

Differential Revision: D24201818

fbshipit-source-id: 691785fbd168c98dff037df1e3fbb2dd5b06d5f6