process link state change for disabled port
Summary:
As titled, if port is disabled, then first delta results in admin state update. This is processed and link is disabled. This triggers another state update where oper state is now marked as down. But because admin state of previous state is also disabled and current state is also disabled, this is not processed. Fixing it.

In summary

- port disable
-  old = admin: up, oper: up, new = admin: down, oper: up          <---- processed
-  old = admin: down, oper: up, new = admin: down, oper: down <--- ignored

Reviewed By: jasmeetbagga

Differential Revision: D24219858

fbshipit-source-id: 6ceab46eefc5b21c44c24ad400f37faa8cd4b8bf